Share feed list sort fields between controllers
The raw SQL sort expressions and the recent-posts query were pasted
verbatim into the user-facing and admin feeds controllers — exactly
the duplication that drifts silently.
